fa24-cs411-team050-Tony-s-Query-Crew/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
appnope==0.1.3
asgiref==3.7.2
asttokens==2.4.1
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
cachetools==5.5.0
certifi==2023.11.17
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloud-sql-python-connector==1.13.0
comm==0.2.1
contourpy==1.2.0
cryptography==43.0.3
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
Django==5.0.2
executing==2.0.1
Flask==3.0.2
fonttools==4.48.1
frozenlist==1.5.0
google-auth==2.36.0
idna==3.10
image==1.5.33
imageio==2.34.0
ipykernel==6.29.0
ipython==8.20.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.0
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
lazy_loader==0.3
lxml==5.3.0
MarkupSafe==2.1.5
matplotlib==3.8.2
matplotlib-inline==0.1.6
multidict==6.1.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.3
packaging==23.2
pandas==2.2.1
parso==0.8.3
patsy==0.5.6
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
propcache==0.2.0
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.17.2
PyMySQL==1.1.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
pyzmq==25.1.2
RateMyProfessorAPI==1.3.6
requests==2.32.3
rsa==4.9
scikit-image==0.22.0
scikit-learn==1.4.2
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
soupsieve==2.6
SQLAlchemy==2.0.36
sqlparse==0.4.4
stack-data==0.6.3
statsmodels==0.14.2
threadpoolctl==3.4.0
tifffile==2024.2.12
tornado==6.4
traitlets==5.14.1
typing_extensions==4.12.2
tzdata==2023.4
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.0.1
yarl==1.17.2
mysql.connector==2.2.9